+++ packages/core/src/header-builder/legacy-csrf-token-header.ts
